# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "KAI API",
        "version": "1.0.0"
    },
    "host": "localhost:8000",
    "basePath": "/",
    "schemes": [
        "http"
    ],
    "paths": {
...
        "/api/stylist/make_style_post": {
            "post": {
                "tags": [
                    "StylePost"
                ],
                "summary": "create style post",
                "description": "create new style post",
                "parameters": [
                    {
                        "name": "desc",
                        "in": "path",
                        "description": "post's description",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "tag",
                        "in": "path",
                        "description": "tags",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "images[]",
                        "in": "path",
                        "description": "images max 10 pieces",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "brand name",
                        "in": "path",
                        "description": "brand name",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "style_name",
                        "in": "path",
                        "description": "style name",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "color",
                        "in": "path",
                        "description": "color",
                        "required": false,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation message ",
                        "schema": {
                            "$ref": "#/definitions/StylePost"
                        }
                    },
                    "422": {
                        "description": "validation error"
                    },
                    "500": {
                        "description": "error something went wrong"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/api/stylist/delete_post/{id}": {
            "get": {
                "tags": [
                    "StylePost"
                ],
                "summary": "delete some post by stylist",
                "description": "delete style post'",
                "responses": {
                    "200": {
                        "description": "Post has been deleted'"
                    },
                    "500": {
                        "description": "error something went wrong"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/api/stylist/show_all_post": {
            "get": {
                "tags": [
                    "StylePost"
                ],
                "summary": "show all post for stylist",
                "description": "delete some style post'",
                "responses": {
                    "200": {
                        "description": "show all post for stylist",
                        "schema": {
                            "$ref": "#/definitions/StylePost"
                        }
                    },
                    "500": {
                        "description": "error something went wrong"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
...
    },
    "definitions": {
...
        "StylePost": {
            "properties": {
                "stylist_id": {
                    "type": "integer"
                },
                "description": {
                    "type": "string"
                },
                "media": {
                    "type": "text"
                },
                "brand_name": {
                    "type": "string"
                },
                "style_name": {
                    "type": "string"
                },
                "tags": {
                    "type": "string"
                },
                "color": {
                    "type": "string"
                }
            }
        },
...
    }
}